refactor(config): split models.py into destination/sync-option modules (#721)
Phase 1 of 2. Split the ~1300-line drt/config/models.py into cohesive modules;
models.py becomes a pure re-export facade with an explicit __all__ (the 72-name
public surface). No class behavior, fields, or validation change — mechanical
move + a describe() base.

Modules:
- base.py            — DescribableConfig (new) + auth, pagination, RetryConfig,
                       LookupConfig, SslConfig, project/source primitives
- destinations_sql.py     — BaseSql family + Snowflake/Databricks/BigQuery/Elasticsearch
- destinations_saas.py    — webhook/CRM/messaging/issue-tracker/email/staged-upload
- destinations_storage.py — S3/GCS/Azure/Parquet/File
- sync_options.py    — RetryConfig consumers, tests, alerts, the DestinationConfig
                       union, SyncOptions/SyncConfig
- models.py          — re-exports all 72 names + __all__

A shared base.py was necessary: RetryConfig is used by ~19 destination configs
while SyncConfig needs the DestinationConfig union — keeping RetryConfig in
sync_options would be a circular import. Lifting the shared primitives into
base.py yields a clean DAG (base -> destinations_* -> sync_options -> models).

DescribableConfig dedups the 26 trivial `f"{type} (detail)"` describe() one-liners
via a _describe_detail hook; 8 connectors with a hardcoded label keep their own.

Verified:
- Whole-package `python -m mypy drt/` clean — 0 no_implicit_reexport errors across
  148 files (the explicit __all__ is what makes the re-exports count under strict
  mypy). Only remaining error is the pre-existing credentials.py:368 (file
  byte-identical to main).
- Structural parity: snapshotted every model's fields/defaults/validators/
  JSON-schema before vs after — 67/67 schemas identical, all validators identical.
- describe() output identical across all 34 destination configs.

Registry-parity comment added at the DestinationConfig union pointing at
tests/unit/test_cli_list_connectors.py::test_DESTINATIONS_matches_registry.
